refactoring and getting posts in radius workin

+++ BlazorWASM-AppLogic/Domain/Models/Coordinate.cs
@@ -2,10 +2,10 @@
 
 public class Coordinate
 {
-    public float latitude{ get; set; }
-    public float longitude{ get; set; }
+    public double latitude{ get; set; }
+    public double longitude{ get; set; }
 
-    public Coordinate(float latitude, float longitude)
+    public Coordinate(double latitude, double longitude)
     {
         this.latitude = latitude;
         this.longitude = longitude;
